Updated NextJs V
Created two diff components in _app.jsx. Functional and Class based, just for testing.
In app.jsx passed props down to component, from getInitialProps, for testing.

Document:

Added metatag in head, consologged props, and added getInitialProps for testing a bug with the Head
not statically being rendered if set in other pages than _document. When getInitialProps & renderPage are in _doc
the Head content in _app.js is not rendered.

Page1, just consologged props and tried to apply Head there too, for testing